Moved Room stuff to data module
