+++ app/src/main/kotlin/io/github/fobo66/wearmmr/App.kt
@@ -17,8 +17,8 @@
 package io.github.fobo66.wearmmr
 
 import android.app.Application
+import io.github.fobo66.data.db.databaseModule
 import io.github.fobo66.wearmmr.api.apiModule
-import io.github.fobo66.wearmmr.db.databaseModule
 import io.github.fobo66.wearmmr.domain.domainModule
 import io.github.fobo66.wearmmr.model.viewModelsModule
 import io.github.fobo66.wearmmr.util.dispatchersModule
