+++ app/src/main/kotlin/io/github/fobo66/wearmmr/model/MainViewModel.kt
@@ -4,7 +4,7 @@ import android.content.SharedPreferences
 import androidx.core.content.edit
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import io.github.fobo66.wearmmr.db.MatchmakingDatabase
+import io.github.fobo66.data.db.MatchmakingDatabase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
