+++ data/src/main/kotlin/io/github/fobo66/data/di/DataModule.kt
@@ -16,7 +16,7 @@ val dataModule = module {
     }
 
     single<PreferenceDataSource> {
-        PreferenceDataSourceImpl()
+        PreferenceDataSourceImpl(get())
     }
 }
